requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
gunicorn==20.1.0
python-dotenv==0.19.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
psycopg2-binary==2.9.9
pytz==2024.1 